# Backend/docs/swagger.yaml
basePath: /
definitions:
...
  domain.ErrorResponse:
    properties:
      error:
        example: Mensaje de error genérico
        type: string
    type: object
...
  domain.Institution:
    properties:
      _id:
        type: string
      color:
        type: string
      name:
        type: string
    type: object
  domain.InstitutionRequest:
    properties:
      color:
        example: '#FF5733'
        type: string
      name:
        example: Hogar de cristo
        type: string
    required:
    - color
    - name
    type: object
...
  domain.Route:
    properties:
      _id:
        type: string
      date_created:
        type: string
      date_finished:
        type: string
      description:
        type: string
      invite_code:
        type: string
      route_leader:
        type: string
      status:
        type: string
      team:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
...
  domain.SuccessResponse:
    properties:
      message:
        example: Mensaje de éxito genérico
        type: string
    type: object
  domain.Usuario:
    properties:
      _id:
        type: string
      completed_routes:
        type: integer
      date_register:
        type: string
      email:
        type: string
      institutionID:
        type: string
      list_routes:
        items:
          $ref: '#/definitions/domain.Route'
        type: array
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Documentación de las APIs del proyecto HDC.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: APIs de HDC
  version: "1.0"
paths:
...
  /export-data/people-helped:
    get:
      consumes:
      - application/json
      description: Genera y descarga un archivo Excel con los datos de todas las personas
        ayudadas registradas en el sistema. Requiere autenticación y rol de administrador.
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: Archivo Excel generado exitosamente
          schema:
            type: file
        "400":
          description: Acceso denegado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Rol no encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Exportar datos de personas ayudadas
      tags:
      - Exportación
...
  /institution:
    get:
      consumes:
      - application/json
      description: Obtiene una lista de todas las instituciones registradas en el
        sistema. Requiere autenticación y rol de administrador.
      produces:
      - application/json
      responses:
        "200":
          description: Lista de instituciones obtenida exitosamente
          schema:
            items:
              $ref: '#/definitions/domain.Institution'
            type: array
        "400":
          description: Error al obtener las instituciones
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Rol no encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obtener todas las instituciones
      tags:
      - Instituciones
    post:
      consumes:
      - application/json
      description: Crea una nueva institución en el sistema. Requiere autenticación
        y rol de administrador.
      parameters:
      - description: Datos de la nueva institución
        in: body
        name: institution
        required: true
        schema:
          $ref: '#/definitions/domain.InstitutionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Institución creada con éxito
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Error al crear la institución
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Rol no encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear nueva institución
      tags:
      - Instituciones
  /institution/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina una institución del sistema mediante su ID. Requiere autenticación
        y rol de administrador.
      parameters:
      - description: ID de la institución
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Institución eliminada correctamente
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Error al eliminar la institución
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Rol no encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Eliminar institución
      tags:
      - Instituciones
    get:
      consumes:
      - application/json
      description: Obtiene una institución específica mediante su ID. Requiere autenticación
        y rol de administrador.
      parameters:
      - description: ID de la institución
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Institución obtenida exitosamente
          schema:
            $ref: '#/definitions/domain.Institution'
        "400":
          description: Error al obtener la institución
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Rol no encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obtener institución por ID
      tags:
      - Instituciones
    put:
      consumes:
      - application/json
      description: Actualiza los datos de una institución existente. Requiere autenticación
        y rol de administrador.
      parameters:
      - description: ID de la institución
        in: path
        name: id
        required: true
        type: string
      - description: Datos actualizados de la institución
        in: body
        name: institution
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Institución actualizada correctamente
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Error al actualizar la institución
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Rol no encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Actualizar institución
      tags:
      - Instituciones
...
  /user:
    get:
      consumes:
      - application/json
      description: Obtiene una lista de todos los usuarios registrados en el sistema.
        Requiere autenticación y rol de administrador.
      produces:
      - application/json
      responses:
        "200":
          description: Lista de usuarios obtenida exitosamente
          schema:
            items:
              $ref: '#/definitions/domain.Usuario'
            type: array
        "400":
          description: Acceso denegado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Rol no encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obtener todos los usuarios
      tags:
      - Usuarios
  /user/{id}:
    get:
      consumes:
      - application/json
      description: Obtiene la información de un usuario específico mediante su ID.
        Requiere autenticación y rol de administrador.
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Usuario obtenido exitosamente
          schema:
            $ref: '#/definitions/domain.Usuario'
        "400":
          description: Acceso denegado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Rol no encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obtener usuario por ID
      tags:
      - Usuarios
  /user/profile:
    get:
      consumes:
      - application/json
      description: Obtiene la información del perfil del usuario autenticado. Requiere
        autenticación.
      produces:
      - application/json
      responses:
        "200":
          description: Perfil de usuario obtenido exitosamente
          schema:
            $ref: '#/definitions/domain.Usuario'
        "400":
          description: Usuario no encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: ID de usuario no encontrado en token
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obtener perfil del usuario
      tags:
      - Usuarios
  /user/public-info/{id}:
    get:
      consumes:
      - application/json
      description: Obtiene la información pública de un usuario específico mediante
        su ID. Requiere autenticación.
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Información pública del usuario obtenida exitosamente
          schema:
            type: object
        "400":
          description: Usuario no encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Token inválido
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obtener información pública del usuario
      tags:
      - Usuarios
  /user/update:
    put:
      consumes:
      - application/json
      description: Actualiza los datos del usuario autenticado. Requiere autenticación.
      parameters:
      - description: Datos a actualizar del usuario
        in: body
        name: user
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Usuario actualizado exitosamente
          schema:
            $ref: '#/definitions/domain.Usuario'
        "400":
          description: Usuario no encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: ID de usuario no encontrado en token
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Actualizar información del usuario
      tags:
      - Usuarios
securityDefinitions:
  BearerAuth:
    description: Se debe proporcionar un token JWT válido con el prefijo "Bearer"
      en el header de autorización para acceder a las rutas protegidas.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"